LayoutTests/transitions/transition-to-from-undefined-expected.txt

CONSOLE MESSAGE: A transition for property max-height could not be found
CONSOLE MESSAGE: Failed to pause 'max-height' transition on element 'test1'
CONSOLE MESSAGE: A transition for property max-height could not be found
CONSOLE MESSAGE: Failed to pause 'max-height' transition on element 'test2'
PASS - "max-height" property for "test1" element at 1s saw something close to: none
PASS - "max-height" property for "test2" element at 1s saw something close to: 100
PASS - "max-height" property for "test3" element at 1s saw something close to: 50
